Cleanup redundant ROM table read functions
