+++ probe-rs/src/architecture/arm/communication_interface.rs
@@ -89,12 +89,30 @@ pub trait ArmProbeInterface: DapAccess + SwdSequence + SwoAccess + Send {
         &mut self,
         access_port: &FullyQualifiedApAddress,
     ) -> Result<Box<dyn ArmMemoryInterface + '_>, ArmError>;
+}
 
-    /// Reads the chip info from the romtable of given debug port.
-    fn read_chip_info_from_rom_table(
-        &mut self,
-        dp: DpAddress,
-    ) -> Result<Option<ArmChipInfo>, ArmError>;
+/// Read chip information from the ROM tables
+pub fn read_chip_info_from_rom_table(
+    probe: &mut dyn ArmProbeInterface,
+    dp: DpAddress,
+) -> Result<Option<ArmChipInfo>, ArmError> {
+    for ap in probe.access_ports(dp)? {
+        if let Ok(mut memory) = probe.memory_interface(&ap) {
+            let base_address = memory.base_address()?;
+            let component = Component::try_parse(&mut *memory, base_address)?;
+
+            if let Component::Class1RomTable(component_id, _) = component {
+                if let Some(jep106) = component_id.peripheral_id().jep106() {
+                    return Ok(Some(ArmChipInfo {
+                        manufacturer: jep106,
+                        part: component_id.peripheral_id().part(),
+                    }));
+                }
+            }
+        }
+    }
+
+    Ok(None)
 }
 
 // TODO: Rename trait!
@@ -256,13 +274,6 @@ impl ArmProbeInterface for ArmCommunicationInterface<Initialized> {
         access_port_address: &FullyQualifiedApAddress,
     ) -> Result<Box<dyn ArmMemoryInterface + '_>, ArmError> {
         ArmCommunicationInterface::memory_interface(self, access_port_address)
-    }
-
-    fn read_chip_info_from_rom_table(
-        &mut self,
-        dp: DpAddress,
-    ) -> Result<Option<ArmChipInfo>, ArmError> {
-        ArmCommunicationInterface::read_chip_info_from_rom_table(self, dp)
     }
 
     fn current_debug_port(&self) -> DpAddress {
